+++ cmsplugin_blocks/admin/feature.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 from django.conf import settings
+from django.utils.translation import gettext_lazy as _
 from django.urls import path
 
 from ..models import Feature
@@ -17,6 +18,7 @@ class FeatureAdmin(admin.ModelAdmin):
     list_display = ("id", "title", "scope", "value", "allowed_plugin_display")
     list_filter = ("scope", "plugins")
     list_editable = ("title", "value")
+    search_fields = ["title", "value"]
 
     @admin.display(empty_value="None")
     def allowed_plugin_display(self, obj):
@@ -26,6 +28,27 @@ class FeatureAdmin(admin.ModelAdmin):
         choices = dict(settings.BLOCKS_FEATURE_PLUGINS)
         names = [str(choices[item]) for item in obj.plugins]
         return ", ".join(names)
+    allowed_plugin_display.short_description = _("Allowed plugins")
+
+    def get_fieldsets(self, request, obj=None):
+        """
+        Define plugin form fieldsets depending features are enabled or not (when there
+        is no defined feature choices).
+        """
+        fieldsets = [
+            (None, {
+                "fields": (
+                    "title",
+                    (
+                        "scope",
+                        "plugins",
+                    ),
+                    "value",
+                ),
+            }),
+        ]
+
+        return tuple(fieldsets)
 
     def get_urls(self):
         """